Feat/locations

+++ src/app/(auth)/registrer/[linkID]/actions.ts
@@ -9,6 +9,7 @@ import { redirect } from "next/navigation"
 import { emailService } from "@/service/email"
 import { EmailTest } from "@/components/email/email-test"
 import { customerService } from "@/service/customer"
+import { locationService } from "@/service/location"
 
 export const signUpAction = publicAction
   .schema(signUpValidation)
@@ -20,7 +21,12 @@ export const signUpAction = publicAction
 
     const isLinkValid = customerService.validateActivationLink(activationLink.inserted)
     if (!isLinkValid) {
-      throw new ActionError("Dit aktiveringslink er udløbet")
+      throw new ActionError("Dit aktiveringslink er ikke længere gyldigt")
+    }
+
+    const existingCustomer = await customerService.getByID(parsedInput.clientID)
+    if (!existingCustomer) {
+      throw new ActionError("Din firmakonto findes ikke")
     }
 
     const existingUser = await userService.getByEmail(parsedInput.email)
@@ -29,19 +35,15 @@ export const signUpAction = publicAction
     }
 
     const newUser = await userService.register({
-      clientID: parsedInput.clientID,
+      customerID: parsedInput.clientID,
       name: parsedInput.name,
       email: parsedInput.email,
       hash: parsedInput.password,
+      role: activationLink.role,
       isActive: true
     })
     if (!newUser) {
       throw new ActionError("Din bruger blev ikke oprettet")
-    }
-
-    const existingCustomer = await customerService.getByID(parsedInput.clientID)
-    if (!existingCustomer) {
-      throw new ActionError("Din firmakonto findes ikke")
     }
 
     if (!existingCustomer.isActive) {
@@ -51,21 +53,29 @@ export const signUpAction = publicAction
       }
     }
 
+    const isAccessAdded = await locationService.addAccess({
+      userID: newUser.id,
+      locationID: activationLink.locationID,
+      customerID: existingCustomer.id,
+      isPrimary: true
+    })
+    if (!isAccessAdded) {
+      throw new ActionError("Der gik noget galt med at give brugeren tilladelse til lokation")
+    }
+
     const isLinkDeleted = await customerService.deleteActivationLink(parsedInput.linkID)
     if (!isLinkDeleted) {
       // NOTE: What to do?
     }
 
-    const newSessionID = await sessionService.create(newUser.id)
+    locationService.setCookie(activationLink.locationID)
+    await sessionService.create(newUser.id)
 
-    const emailError = await emailService.sendOnce(
+    emailService.sendRecursively(
       [parsedInput.email],
       "Velkommen til Nem Lager",
       EmailTest()
     )
-    if (emailError) {
-      throw new ActionError("Kunne ikke sende en velkomst mail")
-    }
 
     redirect("/oversigt")
   })
